Fix regressions that were not passing tests Add wiki syntax support in displayView Fix exception in stats Fix register which should be tested on local database and not main in virtual mode Fix error handling with templates Fix null pointer un UploadAction.java Fix error handling in ViewrevAction.java Fix copydocument mistake Added strings Fix many tests, commenting tests that cannot pass due to setup (check TODO markers) Update test xwiki.cfg and hbm file
git-svn-id: https://svn.xwiki.org/svnroot/xwiki/xwiki/trunk@1002 f329d543-caf0-0310-9063-dda96c69346f

+++ src/main/java/com/xpn/xwiki/store/XWikiHibernateStore.java
@@ -894,11 +894,13 @@ public class XWikiHibernateStore extends XWikiHibernateBaseStore implements XWik
                 cobject.setId(object.getId());
                 if (evict)
                     session.evict(cobject);
-                session.delete(cobject);
+                if (session.contains(cobject))
+                    session.delete(cobject);
             } else {
                 if (evict)
                     session.evict(object);
-                session.delete(object);
+                if (session.contains(object))
+                    session.delete(object);
             }
 
             if (bTransaction) {
@@ -1974,13 +1976,18 @@ public class XWikiHibernateStore extends XWikiHibernateBaseStore implements XWik
         if (context.getWiki().hasDynamicCustomMappings()==false)
            return getSessionFactory();
 
-        List list = searchDocuments(", BaseClass as bclass where bclass.name=doc.fullName and bclass.customMapping is not null",
+        List list;
+        if (useClassesTable(true, context))
+          list = searchDocuments(", BaseClass as bclass where bclass.name=doc.fullName and bclass.customMapping is not null",
+                                    true, false, false, 0, 0, context);
+          list = searchDocuments("",
                                     true, false, false, 0, 0, context);
         boolean result = false;
 
         for (int i=0;i<list.size();i++) {
             XWikiDocument doc = (XWikiDocument)list.get(i);
-            result |= injectCustomMapping(doc.getxWikiClass(), context);
+            if (doc.getxWikiClass().getFieldList().size()>0 )
+             result |= injectCustomMapping(doc.getxWikiClass(), context);
         }
 
         if (result==false)
